ai-sementic-machine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.3.1
faiss-cpu==1.8.0
networkx==3.2.1
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
pydantic==2.6.0
python-dotenv==1.0.1
torch
pytest==7.4.3
httpx==0.25.2
motor==3.3.2
pymongo==4.6.1